Move AgencyBarChart to components/index/cards/charts
